Split the monolithic float8 test file into 3 separate files.
